Move gettext code into i18n module and GettextMixin
