+++ client/plugin.py
@@ -6,6 +6,7 @@ import mad
 from . import jasperpath
 from . import vocabcompiler
 from . import audioengine
+from . import i18n
 
 
 class GenericPlugin(object):
@@ -27,31 +28,13 @@ class AudioEnginePlugin(GenericPlugin, audioengine.AudioEngine):
     pass
 
 
-class SpeechHandlerPlugin(GenericPlugin):
+class SpeechHandlerPlugin(GenericPlugin, i18n.GettextMixin):
     __metaclass__ = abc.ABCMeta
 
     def __init__(self, *args, **kwargs):
         GenericPlugin.__init__(self, *args, **kwargs)
-
-        # Test if language is supported and raise error if not
-        self._get_translations()
-
-    def _get_translations(self):
-        try:
-            language = self.profile['language']
-        except KeyError:
-            language = 'en-US'
-
-        if language not in self._plugin_info.translations:
-            raise ValueError('Unsupported Language!')
-
-        return self._plugin_info.translations[language]
-
-    def gettext(self, *args, **kwargs):
-        return self._get_translations().gettext(*args, **kwargs)
-
-    def ngettext(self, *args, **kwargs):
-        return self._get_translations().ngettext(*args, **kwargs)
+        i18n.GettextMixin.__init__(
+            self, self.info.translations, self.profile)
 
     @abc.abstractmethod
     def get_phrases(self):
