+++ client/pluginstore.py
@@ -1,8 +1,6 @@
 # -*- coding: utf-8 -*-
 import os
-import re
 import logging
-import gettext
 import imp
 import inspect
 import sys
@@ -11,6 +9,7 @@ if sys.version_info < (3, 0):
 else:
     import configparser
 
+from . import i18n
 from . import plugin
 
 
@@ -19,7 +18,7 @@ MANDATORY_OPTIONS = (
     ('Plugin', 'Version'),
     ('Plugin', 'License')
 )
-RE_TRANSLATIONS = re.compile(r'^[a-z]{2}(-[A-Z]{2}){0,1}$')
+
 PLUGIN_INFO_FILENAME = "plugin.info"
 PLUGIN_TRANSLATIONS_DIRNAME = "locale"
 
@@ -45,22 +44,6 @@ def parse_info_file(infofile_path):
 
     logger.debug("Plugin info file '%s' parsed successfully!", infofile_path)
     return cp
-
-
-def parse_translations(translations_path):
-    translations = {}
-    if os.path.isdir(translations_path):
-        for content in os.listdir(translations_path):
-            if not os.path.isdir(os.path.join(translations_path, content)):
-                lang, ext = os.path.splitext(content)
-                if ext == (os.extsep + 'mo') and RE_TRANSLATIONS.match(lang):
-                    with open(os.path.join(translations_path, content),
-                              mode="rb") as f:
-                        translations[lang] = gettext.GNUTranslations(f)
-    if not translations:
-        # Untranslated module, assume hardcoded en-US strings
-        translations['en-US'] = gettext.NullTranslations()
-    return translations
 
 
 def parse_plugin_class(module_name, plugin_directory, superclasses):
@@ -197,7 +180,7 @@ class PluginStore(object):
 
         translations_path = os.path.join(plugin_directory,
                                          self._translations_dirname)
-        translations = parse_translations(translations_path)
+        translations = i18n.parse_translations(translations_path)
 
         module_name = get_module_name(cp.get('Plugin', 'Name'),
                                       cp.get('Plugin', 'Version'))
